+++ test/c/multidef.cpp
@@ -1,0 +1,4 @@
+#include <boost/json/src.hpp>
+
+int main() {
+}